Created new main, moved old to test
I created a new main method to act as the interface to the console as required by the assignment. I moved the old main method to a test file and will transition it to junit tests later. The new main method is still missing implementation for 4 of the required methods, will add later. It also assumes that all values are strings, because I cannot read the type from the console.